pull old_monster_index to own file (untested)
