+++ dadguide/dadguide.py
@@ -20,13 +20,15 @@ from redbot.core import checks, data_manager
 from redbot.core import commands
 
 from .database_manager import *
+from .old_monster_index import MonsterIndex
 
 
 logger = logging.getLogger('red.padbot-cogs.dadguide')
 
 
 def _data_file(file_name: str) -> str:
-    return os.path.join(str(data_manager.cog_data_path(raw_name='dadguide')), file_name)
+    return 'S:\\Documents\\Games\\PAD\\dadguide.sqlite'
+    # return os.path.join(str(data_manager.cog_data_path(raw_name='dadguide')), file_name)
 
 
 CSV_FILE_PATTERN = '{}.csv'
